\name{Plant_Growth_Analysis-package}
\alias{Plant_Growth_Analysis-package}
\alias{Plant Growth Analysis}
\docType{package}
\title{
 Plant growth analysis functions
}
\description{
 This package can calculate groth parameters such as:
 Relative growth rate, Unity leaf rate, shoot - root allometry, etc
}
\details{
\tabular{ll}{
Package: \tab Plant Growth Analysis\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2011-04-04\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
 The most important function in this package is \code{growth_analysis} and
 \code{*_analysis} such as \code{RGR_analysis} which will calculate different
 growth parameters for all the dates supplied.
}
\author{
Luciano Selzer

Maintainer: Luciano Selzer <luciano.selzer@gmail.com>

}
\references{
Hunt, R., Causton, D.R., Shipley, B. & Askew, A.P. (2002) A Modern Tool for Classical Plant Growth Analysis. Annals of Botany, 90, 485-488.
}
\keyword{ package }